fix compiler warnings
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@3562 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/RenderMain/ImageLoader_Shared.cpp
@@ -466,7 +466,7 @@ bool ImageDescriptor::LoadDDSFromFile(FileSpecifier& File, int flags, int actual
 
 		// hehe, resizing to the next power of two could have
 		// introduced another mipmap
-		MipMapCount = 1 + floor(log2(max(Width, Height)));
+		MipMapCount = static_cast<int>(1 + floor(log2(max(Width, Height))));
 		
 		int skip  = 0;
 		if (maxSize) {
